Move generic function to utils instead of base.py and deduplicate cli args logging
